feature: 填空题的试题增删查改

+++ exam-api/src/main/java/com/yf/exam/modules/qu/service/impl/QuServiceImpl.java
@@ -101,7 +101,7 @@ public class QuServiceImpl extends ServiceImpl<QuMapper, Qu> implements QuServic
 
         // 如果更新了图片
         Qu notUpdatedQu = this.getById(qu.getId());
-        if (!notUpdatedQu.getImage().equals(qu.getImage())
+        if (notUpdatedQu != null && !notUpdatedQu.getImage().equals(qu.getImage())
                 && !notUpdatedQu.getImage().equals("")) {
             uploadService.delete(notUpdatedQu.getImage());
         }
@@ -236,36 +236,36 @@ public class QuServiceImpl extends ServiceImpl<QuMapper, Qu> implements QuServic
         List<QuAnswerDTO> answers = qu.getAnswerList();
 
 
-            if (CollectionUtils.isEmpty(answers) && !qu.getQuType().equals(QuType.SAQ)) {
-                throw new ServiceException(1, no + "客观题至少要包含一个备选答案！");
+        if (CollectionUtils.isEmpty(answers) && !qu.getQuType().equals(QuType.SAQ)) {
+            throw new ServiceException(1, no + "客观题至少要包含一个备选答案！");
+        }
+
+
+        int trueCount = 0;
+        for (QuAnswerDTO a : answers) {
+
+            if (a.getIsRight() == null) {
+                throw new ServiceException(1, no + "必须定义选项是否正确项！");
             }
 
-
-            int trueCount = 0;
-            for (QuAnswerDTO a : answers) {
-
-                if (a.getIsRight() == null) {
-                    throw new ServiceException(1, no + "必须定义选项是否正确项！");
-                }
-
-                if (StringUtils.isEmpty(a.getContent())) {
-                    throw new ServiceException(1, no + "选项内容不为空！");
-                }
-
-                if (a.getIsRight()) {
-                    trueCount += 1;
-                }
+            if (StringUtils.isEmpty(a.getContent())) {
+                throw new ServiceException(1, no + "选项内容不为空！");
             }
 
-            if (trueCount == 0 && !qu.getQuType().equals(QuType.SAQ)) {
-                throw new ServiceException(1, no + "至少要包含一个正确项！");
+            if (a.getIsRight()) {
+                trueCount += 1;
             }
+        }
+
+        if (trueCount == 0 && !qu.getQuType().equals(QuType.SAQ)) {
+            throw new ServiceException(1, no + "至少要包含一个正确项！");
+        }
 
 
-            //单选题
-            if (qu.getQuType().equals(QuType.RADIO) && trueCount > 1) {
-                throw new ServiceException(1, no + "单选题不能包含多个正确项！");
-            }
+        //单选题
+        if (qu.getQuType().equals(QuType.RADIO) && trueCount > 1) {
+            throw new ServiceException(1, no + "单选题不能包含多个正确项！");
+        }
 
     }
 }
